# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Prueba Fitpal",
        "description": "Prueba Fitpal",
        "contact": {
            "email": "oscargalvez1992@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/horarios": {
            "get": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Obtener todos los horarios",
                "operationId": "8559ff5b26b44077c955d57b832912a3",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de elementos por página",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "historico",
                        "in": "query",
                        "description": "Incluir horarios históricos",
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Horario"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string"
                                        },
                                        "from": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "last_page_url": {
                                            "type": "string"
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "next_page_url": {
                                            "type": "string"
                                        },
                                        "path": {
                                            "type": "string"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "prev_page_url": {
                                            "type": "string"
                                        },
                                        "to": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Crear un nuevo horario",
                "operationId": "31fcbcf06b9b00ddde6d08ab5a243c4a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Horario"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Horario"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/horarios/{horario}": {
            "get": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Obtener los detalles de un horario",
                "operationId": "7267d98558e1b568981fa21524d07cb6",
                "parameters": [
                    {
                        "name": "horario",
                        "in": "path",
                        "description": "ID del horario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Horario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Horario not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Actualizar un horario existente",
                "operationId": "c4fd412f80498e909f2d934fade55121",
                "parameters": [
                    {
                        "name": "horario",
                        "in": "path",
                        "description": "ID del horario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Horario"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Horario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Horario not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Eliminar un horario existente",
                "operationId": "b7b38cb9eea3dbcae9ce279aba8c99c8",
                "parameters": [
                    {
                        "name": "horario",
                        "in": "path",
                        "description": "ID del horario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Horario not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Clase": {
                "title": "Clase",
                "description": "Clase model",
                "properties": {
                    "id": {
                        "description": "ID of the clase",
                        "type": "integer",
                        "example": "1"
                    },
                    "nombre": {
                        "description": "Nombre of the clase",
                        "type": "string",
                        "example": "Yoga Class"
                    },
                    "descripcion": {
                        "description": "Descripción of the clase",
                        "type": "string",
                        "example": "A relaxing yoga class for all levels"
                    }
                },
                "type": "object"
            },
            "Horario": {
                "title": "Horario",
                "description": "Horario model",
                "properties": {
                    "id": {
                        "description": "ID of the horario",
                        "type": "integer",
                        "example": "1"
                    },
                    "fecha": {
                        "description": "Fecha of the horario",
                        "type": "string",
                        "example": "2023-05-21"
                    },
                    "hora_inicio": {
                        "description": "Hora de inicio del horario",
                        "type": "string",
                        "example": "09:00:00"
                    },
                    "hora_final": {
                        "description": "Hora de finalización del horario",
                        "type": "string",
                        "example": "10:30:00"
                    },
                    "clase_id": {
                        "description": "ID of the related clase",
                        "type": "integer",
                        "example": "1"
                    },
                    "clase": {
                        "$ref": "#/components/schemas/Clase"
                    }
                },
                "type": "object"
            }
        }
    }
}